cloud_storage: refactor into different files
