+++ src/v/cloud_storage/CMakeLists.txt
@@ -24,6 +24,8 @@ v_cc_library(
     recovery_utils.cc
     segment_meta_cstore.cc
     segment_chunk.cc
+    segment_chunk_api.cc
+    segment_chunk_data_source.cc
   DEPS
     Seastar::seastar
     v::bytes
